client.test:to check whether the self.debugdir_tmp_file exist
Sometimes this will lead an error, So before unlink the file
check if it really exists.

Acked-by: Feng Yang <fyang@redhat.com>

+++ client/test.py
@@ -92,7 +92,8 @@ class test(common_test.base_test):
         """
         if self.crash_handling_enabled:
             # Remove the debugdir info file
-            os.unlink(self.debugdir_tmp_file)
+            if os.path.isfile(self.debugdir_tmp_file):
+                os.unlink(self.debugdir_tmp_file)
             # Restore the core pattern backup
             try:
                 utils.open_write_close(self.pattern_file,
